Initial PR for go-for-1.0 (#86)
# Platform Road Map 
The current (approximate) roadmap for the Bosque project is broken down into a series of major components that are being developed in parallel + smaller but high-impact work items. These cover a range of complexity and required skills, from academic level publication results, through non-trivial software engineering lifts, and into smaller tasks that are good starter projects. If you are interested in one of these please find (or open) an issue for more detail.

## Current Work Items 
The current work items are focused on the 1.0 release of the Bosque language and it actively in-progress:
- Base Language design -- mostly workable (in the `assembly.ts` and `body.ts` files) with needs for documentations
- Core Libraries -- not much but lots of opportunity to build out from previous prototype code.
- Parser and Type-Checker -- partially complete with plenty of todos and needs for documentation/tests.
- Basic JS Runtime -- starting...
- Visual Studio Code Integration -- Just a simple syntax highlighter at the moment next steps are to build out a full LSP server.

## Planned Next Steps
The next steps are focused on the 2.0 release of the Bosque language are focused on fulfilling two major promises in the tooling and runtime space:
- Small Model Verifier -- a symbolic checker with guarantees of the decidability of the presence or absence of small inputs that trigger any runtime failure/assertion/pre/post/invariant condition.
- A Ω()-time and O(1)-space runtime and AOT compiler -- a runtime that is designed to be performant, low memory, and predictable in its behavior.